create branch 2.9

+++ src/sbin/photurisd/secrets.h
@@ -1,5 +1,6 @@
+/* $OpenBSD: secrets.h,v 1.4 2001/01/28 22:45:16 niklas Exp $ */
 /*
- * Copyright 1997 Niels Provos <provos@physnet.uni-hamburg.de>
+ * Copyright 1997-2000 Niels Provos <provos@citi.umich.edu>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,7 +28,6 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-/* $Id: secrets.h,v 1.1 1998/11/14 23:37:28 deraadt Exp $ */
 /*
  * secrets.h:
  * prototypes for compute_secrets.c
@@ -38,7 +38,6 @@
  
 #include "state.h" 
 #include "spi.h"
-#include "gmp.h"
  
 #undef EXTERN 
   
@@ -48,7 +47,7 @@
 #define EXTERN extern 
 #endif
 
-EXTERN int compute_shared_secret(struct stateob *, u_int8_t **, u_int16_t *);
+EXTERN int compute_shared_secret(struct stateob *, u_int8_t **, size_t *);
 EXTERN int compute_session_key(struct stateob *st, u_int8_t *key, 
 			       u_int8_t *attribute, int owner, 
 			       u_int16_t *order);
